requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioresponses==0.7.6
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.4.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
propcache==0.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
requests==2.32.3
requests-mock==1.12.1
urllib3==2.2.3
yarl==1.15.1
